Parity errors in NIC are now ignored. Clean up the code to make it easier to understand. Now the logic to error correct the NIC makes more sense.
